Cleanup: move testUtilities out of __tests__
testUtilities are used by other packages, so should be compiled into dist, __tests__ should not be in dist
